extract makeVerb and some spacing functions from buildCommon
