+++ src/functions/symbolsSpacing.js
@@ -9,8 +9,8 @@ import ParseError from "../ParseError";
 defineFunctionBuilders({
     type: "spacing",
     htmlBuilder(group, options) {
-        if (buildCommon.regularSpace.hasOwnProperty(group.text)) {
-            const className = buildCommon.regularSpace[group.text].className || "";
+        if (regularSpace.hasOwnProperty(group.text)) {
+            const className = regularSpace[group.text].className || "";
             // Spaces are generated by adding an actual space. Each of these
             // things has an entry in the symbols table, so these will be turned
             // into appropriate outputs.
@@ -23,10 +23,10 @@ defineFunctionBuilders({
                     [buildCommon.mathsym(group.text, group.mode, options)],
                     options);
             }
-        } else if (buildCommon.cssSpace.hasOwnProperty(group.text)) {
+        } else if (cssSpace.hasOwnProperty(group.text)) {
             // Spaces based on just a CSS class.
             return buildCommon.makeSpan(
-                ["mspace", buildCommon.cssSpace[group.text]],
+                ["mspace", cssSpace[group.text]],
                 [], options);
         } else {
             throw new ParseError(`Unknown type of space "${group.text}"`);
@@ -35,10 +35,10 @@ defineFunctionBuilders({
     mathmlBuilder(group, options) {
         let node;
 
-        if (buildCommon.regularSpace.hasOwnProperty(group.text)) {
+        if (regularSpace.hasOwnProperty(group.text)) {
             node = new mathMLTree.MathNode(
                 "mtext", [new mathMLTree.TextNode("\u00a0")]);
-        } else if (buildCommon.cssSpace.hasOwnProperty(group.text)) {
+        } else if (cssSpace.hasOwnProperty(group.text)) {
             // CSS-based MathML spaces (\nobreak, \allowbreak) are ignored
             return new mathMLTree.MathNode("mspace");
         } else {
@@ -48,3 +48,26 @@ defineFunctionBuilders({
         return node;
     },
 });
+
+// A map of CSS-based spacing functions to their CSS class.
+const cssSpace: {[string]: string} = {
+    "\\nobreak": "nobreak",
+    "\\allowbreak": "allowbreak",
+};
+
+// A lookup table to determine whether a spacing function/symbol should be
+// treated like a regular space character.  If a symbol or command is a key
+// in this table, then it should be a regular space character.  Furthermore,
+// the associated value may have a `className` specifying an extra CSS class
+// to add to the created `span`.
+const regularSpace: {[string]: { className?: string }} = {
+    " ": {},
+    "\\ ": {},
+    "~": {
+        className: "nobreak",
+    },
+    "\\space": {},
+    "\\nobreakspace": {
+        className: "nobreak",
+    },
+};
