+++ src/functions/verb.js
@@ -4,6 +4,8 @@ import buildCommon from "../buildCommon";
 import mathMLTree from "../mathMLTree";
 import ParseError from "../ParseError";
 
+import type {ParseNode} from "../ParseNode";
+
 defineFunction({
     type: "verb",
     names: ["\\verb"],
@@ -20,7 +22,7 @@ defineFunction({
             "\\verb ended by end of line instead of matching delimiter");
     },
     htmlBuilder(group, options) {
-        const text = buildCommon.makeVerb(group, options);
+        const text = makeVerb(group);
         const body = [];
         // \verb enters text mode and therefore is sized like \textstyle
         const newOptions = options.havingStyle(options.style.text());
@@ -38,9 +40,18 @@ defineFunction({
             body, newOptions);
     },
     mathmlBuilder(group, options) {
-        const text = new mathMLTree.TextNode(buildCommon.makeVerb(group, options));
+        const text = new mathMLTree.TextNode(makeVerb(group));
         const node = new mathMLTree.MathNode("mtext", [text]);
         node.setAttribute("mathvariant", "monospace");
         return node;
     },
 });
+
+/**
+ * Converts verb group into body string.
+ *
+ * \verb* replaces each space with an open box \u2423
+ * \verb replaces each space with a no-break space \xA0
+ */
+const makeVerb = (group: ParseNode<"verb">): string =>
+    group.body.replace(/ /g, group.star ? '\u2423' : '\xA0');
